% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{add_val_buffer}
\alias{add_val_buffer}
\title{Add some amount of buffer to a value.
Used to calculate the min/max amounts
for ylimits}
\usage{
add_val_buffer(val, buffer_pct = 0.01, ceilingBuf = TRUE, asInteger = FALSE)
}
\arguments{
\item{val}{A numeric type.}

\item{buffer_pct}{The proportion of val to add/subtract to/from val.}

\item{ceilingBuf}{If true, add buffer_pct to val. Otherwise, subtract buffer_pct from val.}

\item{asInteger}{If true, return the nearest ceiling/floor integer to the buffered value.}
}
\value{
A numeric value (v +- buffer_pct * val)
}
\description{
Add some amount of buffer to a value.
Used to calculate the min/max amounts
for ylimits
}
