% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{calculate_smart_ylimits}
\alias{calculate_smart_ylimits}
\title{Returns a set of ylimits trimmed to fit the input dataframe.}
\usage{
calculate_smart_ylimits(x, fitMinimum = FALSE)
}
\arguments{
\item{x}{A TidySig tibble}

\item{fitMinimum}{fit a minimum value instead of defaulting to zero.}
}
\value{
A vector of two elements containing calculated ylimits.
}
\description{
Returns a set of ylimits trimmed to fit the input dataframe.
}
