% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{combine_context_change_cols}
\alias{combine_context_change_cols}
\title{A helper function to 
1. Properly rename columns in SigProfiler inputs
2. Transform the Context/Change columns to the proper format.}
\usage{
combine_context_change_cols(x)
}
\arguments{
\item{x}{A dataframe of Sigprofiler signatures / counts}
}
\value{
a tibble, with the new columns "Change" and "Context"
}
\description{
A helper function to 
1. Properly rename columns in SigProfiler inputs
2. Transform the Context/Change columns to the proper format.
}
