% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{convert_pcawg_to_sigprofiler_SBS96}
\alias{convert_pcawg_to_sigprofiler_SBS96}
\title{Convert a pcawg-formatted dataframe to modern sigprofiler SBS96.
PCAWG's sigprofiler formatting is different from that of
the current version, so we munge the column names.}
\usage{
convert_pcawg_to_sigprofiler_SBS96(x)
}
\arguments{
\item{x}{a dataframe from PCAWG's sigprofiler examples (usually read in with readr::read_csv)}
}
\value{
A tibble with the column names conforming to SigProfiler's expected column names.
}
\description{
Convert a pcawg-formatted dataframe to modern sigprofiler SBS96.
PCAWG's sigprofiler formatting is different from that of
the current version, so we munge the column names.
}
