% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{determine_context}
\alias{determine_context}
\title{A helper function to determine the context type of a dataframe}
\usage{
determine_context(x)
}
\arguments{
\item{x}{A TidySig dataframe}
}
\value{
A string indicating the mutational signature context of x
}
\description{
A helper function to determine the context type of a dataframe
}
