% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_enforced_signatures.R
\name{drop_enforced_signatures}
\alias{drop_enforced_signatures}
\title{Remove enforced signatures from a dataframe}
\usage{
drop_enforced_signatures(x)
}
\arguments{
\item{x}{A Tidy-formatted dataframe / tibble of signatures}
}
\value{
A dataframe with all rows from enforced signatures removed
}
\description{
Remove enforced signatures from a dataframe
}
