% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{export_to_signatureanalyzer_word_SBS96}
\alias{export_to_signatureanalyzer_word_SBS96}
\title{Transform a tibble in TidySig format
to SigantureAnalyzer's word format.}
\usage{
export_to_signatureanalyzer_word_SBS96(x)
}
\arguments{
\item{x}{a Tidy-formatted tibble of signatures}
}
\value{
A tibble in SignatureAnalyzer's word format
}
\description{
Transform a tibble in TidySig format
to SigantureAnalyzer's word format.
}
