% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{export_to_sigprofiler_SBS96}
\alias{export_to_sigprofiler_SBS96}
\title{Convert a Tidy-formatted tibble to the SigProfiler matrix format}
\usage{
export_to_sigprofiler_SBS96(x)
}
\arguments{
\item{x}{a Tidy-formatted tibble of Sigprofiler signatures / counts}
}
\value{
A tibble in the expected SigProfiler format
}
\description{
Convert a Tidy-formatted tibble to the SigProfiler matrix format
}
