% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{generate_plot}
\alias{generate_plot}
\title{A helper function for plotting ID83 / SBS96 plots.}
\usage{
generate_plot(
  x,
  label = "Signature",
  title = "",
  xlabel = NULL,
  ylabel = "Count",
  usePercent = TRUE,
  ylimits = NULL,
  contextType = "SBS96"
)
}
\arguments{
\item{x}{A TidySig dataframe/tibble}

\item{label}{The right-side (i.e., facet) label.
Usually "Signature" or "Sample" or a sample ID.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{A y-axis label}

\item{usePercent}{Use percent scales (rather than counts)}

\item{ylimits}{Use custom ylimits (useful for normalizing the views of multiple signatures).
Takes a numeric vector length-two.}

\item{contextType}{An argument for determining the context, either ID83 or SBS96}
}
\value{
a ggplot2 object
}
\description{
A helper function for plotting ID83 / SBS96 plots.
}
