% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{get_minimum_amount}
\alias{get_minimum_amount}
\title{Retrieve the minimum value
of a signature or activity dataframe.
This function can be used to set intelligent y-limits
in plotting functions.}
\usage{
get_minimum_amount(x)
}
\arguments{
\item{x}{a TidySig tibble}
}
\value{
A numeric with the lowest Amount in the dataframe
}
\description{
Retrieve the minimum value
of a signature or activity dataframe.
This function can be used to set intelligent y-limits
in plotting functions.
}
