% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysig_contexts_and_palettes.R
\docType{data}
\name{id_83_colors}
\alias{id_83_colors}
\title{A color palette for ID83 changes. Matches that of SigProfilerPlotting}
\format{
An object of class \code{character} of length 16.
}
\usage{
id_83_colors
}
\description{
A color palette for ID83 changes. Matches that of SigProfilerPlotting
}
\keyword{datasets}
