% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_signatureanalyzer.R
\name{load_pandas_hdf5_data}
\alias{load_pandas_hdf5_data}
\title{Loads a Pandas data frame, restricting to a top level path restrictToPath if desired}
\usage{
load_pandas_hdf5_data(finame, restrictToPath = "")
}
\arguments{
\item{finame}{The name of the input HDF5 file}

\item{restrictToPath}{a top-level path in the HDF5 file.}
}
\description{
Loads a Pandas data frame, restricting to a top level path restrictToPath if desired
}
