% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Transform per-signature feature counts to proportions}
\usage{
normalize_counts(x)
}
\arguments{
\item{x}{a tibble or dataframe (TidySig format) of counts}
}
\value{
a tibble with the Amount column containing proportions per-signature.
}
\description{
Transform per-signature feature counts to proportions
}
