% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_sample_counts}
\alias{normalize_sample_counts}
\title{Convert per-sample counts to proportions}
\usage{
normalize_sample_counts(x)
}
\arguments{
\item{x}{a TidySig tibble}
}
\value{
A TidySig tibble with the Amount column now containing a proportion.
}
\description{
Convert per-sample counts to proportions
}
