% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_sample_summary.R
\name{per_sample_counts_summary}
\alias{per_sample_counts_summary}
\title{Generate a sample-level summary of Signature counts}
\usage{
per_sample_counts_summary(x, label = "Sigprofiler")
}
\arguments{
\item{x}{A TidySig tibble}

\item{label}{An optional label to append to the output column name(s)}
}
\value{
A tibble, with new columns (one per signature) listing per-sample counts of that signature.
}
\description{
Generate a sample-level summary of Signature counts
}
