% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_context}
\alias{plot_context}
\title{A helper function to plot mutational contexts,
which are generally counts (in contrast to signatures,
which are generally proportions).}
\usage{
plot_context(
  x,
  label = "Signature",
  title = "",
  xlabel = "Motif Length or Count",
  ylabel = "Count",
  usePercent = FALSE,
  ylimits = NULL,
  countsAsProportions = FALSE,
  separatePlots = FALSE,
  contextType = contextType
)
}
\arguments{
\item{x}{A TidySig dataframe/tibble}

\item{label}{The right-side (i.e., facet) label.
Usually "Signature" or "Sample" or a sample ID.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{A y-axis label}

\item{usePercent}{Use percent scales (rather than counts)}

\item{ylimits}{Use custom ylimits (useful for normalizing the views of multiple signatures).
Takes a numeric vector length-two.}

\item{countsAsProportions}{Use proportions, rather than raw counts.}

\item{separatePlots}{Produce separate plots, one per Signature}

\item{contextType}{An argument for determining the context, either ID83 or SBS96}
}
\description{
A helper function to plot mutational contexts,
which are generally counts (in contrast to signatures,
which are generally proportions).
}
