% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ID83_signature.R
\name{plot_ID83_signature}
\alias{plot_ID83_signature}
\title{Plot an ID83 signature ro series of signatures}
\usage{
plot_ID83_signature(
  x,
  label = "Signature",
  title = "",
  xlabel = "Motif Length or Count",
  ylabel = "Count",
  usePercent = TRUE,
  ylimits = NULL,
  countsAsProportions = FALSE
)
}
\arguments{
\item{x}{A TidySig dataframe/tibble}

\item{label}{The right-side (i.e., facet) label.
Usually "Signature" or "Sample" or a sample ID.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{A y-axis label}

\item{usePercent}{Use percent scales (rather than counts)}

\item{ylimits}{Use custom ylimits (useful for normalizing the views of multiple signatures)}

\item{countsAsProportions}{Convert the input data (in counts) to per-signature proportions}
}
\value{
a ggplot2 object
}
\description{
Plot an ID83 signature ro series of signatures
}
