% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SBS96_signature.R
\name{plot_SBS96_signature}
\alias{plot_SBS96_signature}
\title{Plot an SBS96 signature or series of signatures}
\usage{
plot_SBS96_signature(
  x,
  label = "Signature",
  title = NULL,
  xlabel = "Base Context",
  ylabel = "Count",
  ylimits = NULL,
  usePercent = FALSE,
  countsAsProportions = FALSE,
  facetCondition = NULL
)
}
\arguments{
\item{x}{A TidySig dataframe/tibble}

\item{label}{The right-side (i.e., facet) label.
Usually "Signature" or "Sample" or a sample ID.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{A y-axis label}

\item{ylimits}{Use custom ylimits (useful for normalizing the views of multiple signatures).
Takes a numeric vector length-two OR a string "smart" to indicate that consistent y-limits should
be automatically fit from the values in x.}

\item{usePercent}{Use percent scales (rather than counts)}

\item{countsAsProportions}{Convert the input data (in counts) to per-signature proportions}

\item{facetCondition}{a condition to generate facet columns.}
}
\value{
a ggplot2 object
}
\description{
Plot an SBS96 signature or series of signatures
}
