% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activities.R
\name{plot_signature_activities}
\alias{plot_signature_activities}
\title{Plot the activities of signature(s)}
\usage{
plot_signature_activities(
  x,
  title = NULL,
  xlabel = "Samples",
  ylabel = "Mutation Counts",
  ylimits = NULL,
  usePercent = FALSE,
  countsAsProportions = FALSE,
  showSampleNames = FALSE,
  orderByMutationCount = TRUE,
  facetGroupVariable = NULL
)
}
\arguments{
\item{x}{A TidySig dataframe/tibble}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{A y-axis label}

\item{ylimits}{Use custom ylimits (useful for normalizing the views of multiple signatures)}

\item{usePercent}{Use percent scales (rather than counts)}

\item{countsAsProportions}{Convert the input data (in counts) to per-signature proportions}

\item{showSampleNames}{display sample names underneath the plot}

\item{orderByMutationCount}{sort the data by descending number of mutations}

\item{facetGroupVariable}{a variable to facet the data on (e.g., an exposure condition).}
}
\value{
a ggplot2 object
}
\description{
Plot the activities of signature(s)
}
