% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{read_pcawg_SBS96_csv}
\alias{read_pcawg_SBS96_csv}
\title{Read in a PCAWG-format CSV file}
\usage{
read_pcawg_SBS96_csv(filename)
}
\arguments{
\item{filename}{The name of a CSV input file}
}
\value{
a TidySig tibble
}
\description{
Read in a PCAWG-format CSV file
}
