% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_signature_plot.R
\name{save_signature_plot}
\alias{save_signature_plot}
\title{Save a plot for a single mutational signature}
\usage{
save_signature_plot(x, filename, plot_height = 4)
}
\arguments{
\item{x}{A ggplot object}

\item{filename}{The name of the PDF to generate (must end in .pdf).}

\item{plot_height}{The height of the plot to generate. All plots have
a 1:3 aspect ratio.}
}
\description{
This function saves a plot object x (usualy from plot_*_signature)
to a PDF named filename.
}
