% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysig_contexts_and_palettes.R
\docType{data}
\name{sbs_96_changes}
\alias{sbs_96_changes}
\title{A list of the six possible (normalized) SBS96 mutational changes}
\format{
An object of class \code{character} of length 6.
}
\usage{
sbs_96_changes
}
\description{
A list of the six possible (normalized) SBS96 mutational changes
}
\keyword{datasets}
