% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysig_contexts_and_palettes.R
\docType{data}
\name{sbs_96_changes_colors}
\alias{sbs_96_changes_colors}
\title{A color palette for SBS96 changes. Matches that of SigProfilerPlotting}
\format{
An object of class \code{character} of length 6.
}
\usage{
sbs_96_changes_colors
}
\description{
A color palette for SBS96 changes. Matches that of SigProfilerPlotting
}
\keyword{datasets}
