% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysig_contexts_and_palettes.R
\docType{data}
\name{sbs_96_contexts}
\alias{sbs_96_contexts}
\title{A list of the SBS96 mutational contexts}
\format{
An object of class \code{character} of length 96.
}
\usage{
sbs_96_contexts
}
\description{
A list of the SBS96 mutational contexts
}
\keyword{datasets}
