% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activities.R
\name{signature_activity_figure}
\alias{signature_activity_figure}
\title{Plot the activities of signature(s) as both counts and proportions}
\usage{
signature_activity_figure(
  x,
  legendPosition = "bottom",
  countYlim = NULL,
  propYlim = NULL
)
}
\arguments{
\item{x}{A TidySig tibble}

\item{legendPosition}{One of ("top", "bottom", "right" or "left") for legend placement}

\item{countYlim}{custom y-limits for the count plot}

\item{propYlim}{custom y-limits for the proportions plot}
}
\value{
a ggplot2 object of two plots, one of counts and one of proportions
}
\description{
Plot the activities of signature(s) as both counts and proportions
}
