% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{tidy_sigprof_activities}
\alias{tidy_sigprof_activities}
\title{Transform sigprofiler activities into TidySig format}
\usage{
tidy_sigprof_activities(x)
}
\arguments{
\item{x}{a Sigprofiler dataframe}
}
\value{
a TidySig tibble
}
\description{
Transform sigprofiler activities into TidySig format
}
