% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{tidy_sigprof_ID83_df}
\alias{tidy_sigprof_ID83_df}
\title{A wrapper around transform_sigprofiler_df for legacy compatibility (ID83 contexts)}
\usage{
tidy_sigprof_ID83_df(x)
}
\arguments{
\item{x}{A dataframe of sigprofiler counts / signatures}
}
\value{
A Tidy-formatted tibble of signature amounts.
}
\description{
A wrapper around transform_sigprofiler_df for legacy compatibility (ID83 contexts)
}
