% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{tidy_sigprof_id83_probabilities}
\alias{tidy_sigprof_id83_probabilities}
\title{A fuction for formatting the signature probability files of SigProfiler (ID83 context)}
\usage{
tidy_sigprof_id83_probabilities(x)
}
\arguments{
\item{x}{a dataframe of SigProfiler ID83 probabilities per signatures}
}
\value{
A Tidy-formatted tibble of probabilities.
}
\description{
A fuction for formatting the signature probability files of SigProfiler (ID83 context)
}
