% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{tidy_sigprof_SBS96_df}
\alias{tidy_sigprof_SBS96_df}
\title{A wrapper around transform_sigprofiler_df for legacy compatibility}
\usage{
tidy_sigprof_SBS96_df(x)
}
\arguments{
\item{x}{a SigProfiler signature dataframe}
}
\value{
A Tidy-formatted tibble of signature amounts.
}
\description{
A wrapper around transform_sigprofiler_df for legacy compatibility
}
