% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{tidy_sigprof_sbs96_probabilities}
\alias{tidy_sigprof_sbs96_probabilities}
\title{A fuction for formatting the signature probability files of SigProfiler}
\usage{
tidy_sigprof_sbs96_probabilities(x)
}
\arguments{
\item{x}{a dataframe of SigProfiler SBS96 probabilities per signatures}
}
\value{
A Tidy-formatted tibble of probabilities.
}
\description{
A fuction for formatting the signature probability files of SigProfiler
}
