% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysig.R
\docType{package}
\name{tidysig}
\alias{tidysig}
\title{\code{tidysig} package}
\description{
tidysig: A tidyverse-style package for plotting mutational signatures.
}
\details{
See the README on
\href{https://github.com/edawson/tidysig/README.md}{GitHub}
}
