% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{transform_sigprofiler_df}
\alias{transform_sigprofiler_df}
\title{Transform a Sigprofiler dataframe (of counts, percentages,
signatures, or activities) into a Tidy format.}
\usage{
transform_sigprofiler_df(x)
}
\arguments{
\item{x}{A dataframe from the SigProfiler output (usually read in with readr::read_tsv)}
}
\value{
A Tidy-formatted tibble of signatures, activities, or counts.
}
\description{
Transform a Sigprofiler dataframe (of counts, percentages,
signatures, or activities) into a Tidy format.
}
