% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sigprofiler.R
\name{transpose_sigprofiler_df}
\alias{transpose_sigprofiler_df}
\title{Transpose a SigProfiler dataframe}
\usage{
transpose_sigprofiler_df(x)
}
\arguments{
\item{x}{A dataframe produced by reading in a SigProfiler matrix (usually with readr::read_tsv)}
}
\value{
A tibble with Context/Changes as rows and the Sample/Signature as rows.
}
\description{
Transpose a SigProfiler dataframe
}
